move meta tag script service methods to new MetaTagScriptService
